Feat: add jwt token and authorization

+++ internal/repository/todo_repository.go
@@ -9,26 +9,27 @@ import (
 	"github.com/jackc/pgx/v5/pgxpool"
 )
 
-func CreateTodo(pool *pgxpool.Pool, title string, completed bool) (*models.Todo, error){
+func CreateTodo(pool *pgxpool.Pool, title string, completed bool, userID string) (*models.Todo, error){
 	var ctx context.Context
 	var cancel context.CancelFunc
 	ctx, cancel = context.WithTimeout(context.Background(), 5*time.Second)
 	defer cancel()
 
 	var query string = `
-		INSERT INTO todos (title, completed)
-		VALUES ($1, $2)
-		RETURNING id, title, completed, created_at, updated_at
+		INSERT INTO todos (title, completed, user_id)
+		VALUES ($1, $2, $3)
+		RETURNING id, title, completed, created_at, updated_at, user_id
 	`
 
 	var todo models.Todo
 	
-	var err = pool.QueryRow(ctx, query, title, completed).Scan(
+	var err = pool.QueryRow(ctx, query, title, completed, userID).Scan(
 		&todo.ID,
 		&todo.Title,
 		&todo.Completed,
 		&todo.CreatedAt,
 		&todo.UpdatedAt,
+		&todo.UserID,
 	)
 
 	if err != nil {
@@ -38,7 +39,7 @@ func CreateTodo(pool *pgxpool.Pool, title string, completed bool) (*models.Todo,
 	return &todo, err
 }
 
-func GetAllTodos (pool *pgxpool.Pool) ([]models.Todo, error) {
+func GetAllTodos (pool *pgxpool.Pool, userID string) ([]models.Todo, error) {
 	var ctx context.Context
 	var cancel context.CancelFunc
 	ctx, cancel = context.WithTimeout(context.Background(), 5 *time.Second)
@@ -46,9 +47,10 @@ func GetAllTodos (pool *pgxpool.Pool) ([]models.Todo, error) {
 
 	var query string = `
 		SELECT * FROM todos 
+		WHERE user_id = $1
 		ORDER BY created_at 
 	`
-	var rows, err = pool.Query(ctx, query)
+	var rows, err = pool.Query(ctx, query, userID)
 	
 	if err != nil {
 		return nil, err
@@ -65,6 +67,7 @@ func GetAllTodos (pool *pgxpool.Pool) ([]models.Todo, error) {
 			&todo.Completed,
 			&todo.CreatedAt,
 			&todo.UpdatedAt,
+			&todo.UserID,
 		)
 		if err != nil {
 			return nil, err
@@ -79,22 +82,25 @@ func GetAllTodos (pool *pgxpool.Pool) ([]models.Todo, error) {
 	return todos, nil
 }
 
-func GetTodoByID (pool *pgxpool.Pool, id int) (*models.Todo, error) {
+func GetTodoByID (pool *pgxpool.Pool, id int, userID string) (*models.Todo, error) {
 	var ctx context.Context
 	var cancel context.CancelFunc
 	ctx, cancel = context.WithTimeout(context.Background(), 5*time.Second)
 	defer cancel()
 
 	var query string = `
-		SELECT id, title, completed, created_at, updated_at FROM todos WHERE id = $1
+		SELECT id, title, completed, created_at, updated_at, user_id 
+		FROM todos 
+		WHERE id = $1 AND user_id = $2
 	`
 	var todo models.Todo
-	var err = pool.QueryRow(ctx, query, id).Scan(
+	var err = pool.QueryRow(ctx, query, id, userID).Scan(
 		&todo.ID,
 		&todo.Title,
 		&todo.Completed,
 		&todo.CreatedAt,
 		&todo.UpdatedAt,
+		&todo.UserID,
 	)
 	if err != nil {
 		return nil, err
@@ -103,7 +109,7 @@ func GetTodoByID (pool *pgxpool.Pool, id int) (*models.Todo, error) {
 	return &todo, nil
 }
 
-func UpdateTodoByID(pool *pgxpool.Pool, id int, title string, completed bool) (*models.Todo, error){
+func UpdateTodoByID(pool *pgxpool.Pool, id int, title string, completed bool, userID string) (*models.Todo, error){
 	var ctx context.Context
 	var cancel context.CancelFunc
 
@@ -113,18 +119,19 @@ func UpdateTodoByID(pool *pgxpool.Pool, id int, title string, completed bool) (*
 	var query string = `
 		UPDATE todos
 		SET title = $1, completed = $2, updated_at = CURRENT_TIMESTAMP
-		WHERE id = $3
-		RETURNING id, title, completed, created_at, updated_at
+		WHERE id = $3 AND user_id = $4
+		RETURNING id, title, completed, created_at, updated_at, user_id
 	`
 
 	var todo models.Todo
 
-	var err error = pool.QueryRow(ctx, query, title, completed, id).Scan(
+	var err error = pool.QueryRow(ctx, query, title, completed, id, userID).Scan(
 		&todo.ID,
 		&todo.Title,
 		&todo.Completed,
 		&todo.CreatedAt,
 		&todo.UpdatedAt,
+		&todo.UserID,
 	)
 
 	if (err !=  nil){
@@ -134,7 +141,7 @@ func UpdateTodoByID(pool *pgxpool.Pool, id int, title string, completed bool) (*
 	return &todo, nil
 }
 
-func DeleteTodoByID(Pool *pgxpool.Pool, id int) error {
+func DeleteTodoByID(Pool *pgxpool.Pool, id int, userID string) error {
 	var ctx context.Context
 	var cancel context.CancelFunc
 
@@ -142,9 +149,9 @@ func DeleteTodoByID(Pool *pgxpool.Pool, id int) error {
 	defer cancel()
 
 	var query string = `
-		DELETE FROM todos WHERE id = $1
+		DELETE FROM todos WHERE id = $1 AND user_id = $2
 	`
-	var commandTag, err = Pool.Exec(ctx, query, id)
+	var commandTag, err = Pool.Exec(ctx, query, id, userID)
 
 	if err != nil {
 		return err
